File choosing assistance
When choosing files, what used to happen is if I, say, had a file named "A" and file named "a" they would not appear together because that is *technically* what alphabetical order says. Other than that, it would work just fine in "alphabetical" order, ABCDEFabcdef.  
However, in Mindustry font, caps and lower looks the same. So it looks really weird: ABCDEFABCDEF.  
  
This change makes it so it doesn't matter the caps/un-caps, it will sort the above files as AABBCCDDEEFF.  
(The slight problem comes when someone has two files with the same name but different caps at the start, but that seems like a less common problem, I feel.)