Add `isConservative` status to `PowerGenerator` blocks
This allows us to track which blocks have the capability to conserve
their input fuel items and which do not. This is set to true for
`ItemLiquidGenerator`, so that an initial implementation can be done
based on only this type of block conserving the input fuel. The
intention is that this behaviour could be expanded to other types of
block later, if this is desired.