Implement conservative generators
When the power needed plus the power put into batteries is less than the
total generation capacity, conservative generators will conserve a
portion of their input items proportional to the percentage of optional
conservative power generation, which is unused. The conservation is
carried out based on extending the amount of time until the generators
will consume an additional item of fuel. So for example if the power
needed is 50% of the production capacity which is entirely conservative
then the generators will take 2x as long between consuming items of
fuel.