Bullet spawn shifting
If for some reason someone makes a turret that doesn't have its barrel at the top of its sprite, you could shift it to the side, or forwards and backwards. Works for moving both bullets of an alternating turret, and dunno why anyone would need it but there's yRand as well.