Removed phase radius display *(see description)*
Phased radius has been proposed in numerous PRs, but I don't know how to handle it properly yet. New players may be confused about the extra circle/hexagon indicator around blocks - it's not clear what it means. An extra radius easy to implement, so I'll leave it out of this PR and handle it later.