Update bundle_ro.properties
This is not ready, hence will be a draft. ea377c6f1608aa17b1167b949aec7a44088d5333 still unfinished. Do not merge as-is. It is unfinished and untested.

This is a rushed submission. I did not want to lose my progress; don't have time to finish it right now.