Completed italian bundle
Now fixes, improvements and **names of units** are applied (_early version, they'll be improved_) thanks to the collaborations with italian persons who compared thoughts, opinions and ideas. I want to give special thanks to @GioIacca9 for his suggestions and his bundle that I review and I found some mistakes but nothing dramatic. We hope we will able to work together for a better translation. 
@Anuken, this bundle is ready to be *merged*.