Add a clear button, to clear generation.
This is because its very annoying to have to spam click if you want to clear the generation.

Also, adds the download icon to the "Reset to Default" button since it works with that i think and the other buttons have icons in the menu.

Also resized the icons to make the text less cramped looking.