Increase Surge Swarmer Ammo Multipler
Increase Surge alloy Ammo Multiplier from x2 to x5
Increase Pyratite Ammo Multiplier from x2 to x3
x2 Is far too low for a small and fast firing turret like swarmer, especially for surge alloy, which is very expensive to make. This is also different from the blast multiplier, which is at x4.
This low multiplier is actively hindering the viability of swarmer, as currently blast compound is the only viable ammunition.
This cosensus has been reach over #balancing on the Mindustry Discord and over issue #1180. Although the surge multiplier was the heat of the discussion the Pyratite multipler is also too low, considering the multiplier for metaglass on cyclones.