Change numeral units to support various numbering systems
- Deprecate `unit.thousands`, `unit.millions` and `unit.billions`.
- Introduce `unit.numeric`.

# `unit.numeric`

This is comma-separated list of numeral units. Each part is meaning 1, 10, 100, 1000 and so on.

For example: the position of `unit.billions` is 10th, because of the value 1,000,000,000 has 10 digits. So `unit.numeric` is `,,,,,,,,b`.

The positions of `unit.thousands` and `unit.millions` are 4th and 7th. So combined `unit.numeric` becomes `,,,k,,,mil,,,b`.