I've translated this list up to 204.
I've translated this list up to 204, for the most part, spell checks and small changes. I'll continue this, though I'm both a bit tired of it and annoyed of modern computer/internet based terms that I have to translate without sounding like I'm in the middle of the 13th century nor like I'm illiterate.