Update bundle_id_ID.properties
Well, Here i did a lot and Huge improve, probably.
In Indonesian Language, do not translate things that are not supposed to translated. Cuz it's reallly doesn't make sense at all. i glad my friends didn't make a criticism on Google App or even on Steam. Well for The Standard Word, The Name of Turrets, Core Bastion became 'Baluarti' huh? (it's Spannish dude), including Green Liquid, The Arkycite, is not supposed to be translating. Real Reason their are not existing. And what on earth, there so much much much Weird Translation in Indonesian Language. Sucks dude, to be honest. Even My friend tell me, "Weird turret name, it doesn't same with Original Name"
Well that's enough. Please just Admit The Requiest Directly to The Server, I knew at some point it going to be Draft Request, but, JUST ADMIT IT IN THE UPDATE. I'm just wanna make Indonesian Language in this game better, and not to make it bad, especially with that Spectre Turret became 'Momok'
Really Weird. Well see ya!