Added elapsed time counter
With this update, a simple boolean-enabled elapsed time feature is added. All this does is show how much time elapses between every frame while updating it. This counter outputs that every second via logging to console. This update can and should be ignored, as it is only being done for a pull request. In that pull request, we are to detail said elapsed time as well as frame rate measured by running 3  relatively large-scale games.

I was tasked to benchmark Mindustry on games with thousands of blocks. I found three relatively large games with hundreds of nodes, potentially thousands. Each of these games were joined twice, and played for 10 minutes for accurate data. For the first map, on the first join, I got a frame rate of 20, and an elapsed time of 31830 milliseconds averaged over a minute. For the second join, it was 30 frames, with an elapsed time per frame of 32071 milliseconds averaged over a minute.

For the second map, on first join, the frame rate was 26, and its elapsed time was 26409 milliseconds. With respect to its second join, the frame rate was 19, and the elapsed time was 32806 milliseconds.

For the third and final map, after first joining it, I calculated a frame rate of 13 per second, and an elapsed time of 32269 milliseconds between frame update, averaged over a minute. After joining it a second time, the frame rate was 18, and the elapsed time was 26956 milliseconds per minute.

Per second, all of these maps and sessions had a minimum of 500 milliseconds between frames every second, and a maximum of 700 milliseconds between frames.