Update bundle_hu.properties
I tried to rewrite the question sentences, but eventually I realised that they should not be rewritten.

Two reasons:

1) if someone starts translating after me, they won't understand what happened to the question sentences and will see it as a mistake and will try to correct it.
2) the sentences became longer after the rewrite, and nobody likes to read a lot while playing.

Another solution:

There is one solution to fix the question sentences and that is to introduce a Yes answer in the answer options instead of OK. However, this only applies to sentences ending in ?. The other sentences should remain with the OK answer option.

Yes I know, there is no time for this and it is not such a fatal mistake. Unfortunately I can't do it because I'm a programming noob. It will be done when it's done, it's not that important.

-----------This Pull Request-----------

This PR includes recovery of accidentally deleted items, typing corrections, spelling corrections, simplification of question sentences, and translation of the new sentence.