Team Color refactor
De-Finalizes Team colors (Team.color, Team.palette, Team.palette1), and implements setPalette() with two variants; either generating a palette from a single color (as the green & blue teams do), or filling one out from three manually provided colors.  This refactors the  Team constructors to use setPalette as well.

This is realistically only going to matter to like 3 people in the community (and all of them are me) by allowing team color editing in mods.  If there's a performant way to alter colors at runtime, that could be cool for scripted maps or similar, but I otherwise expect this to mostly be used for some simple recoloring mods/use in full conversions to freshen things up.