Allowing multiple planets to have the same tech tree (#11017)
* Lists all planets having the same TechTree associated with them.

As of right now, this problem arizes only with modded games. If you add planets in a specific order (e.g. StarA, PlanetA, PlanetB) and attach same tech tree to all of them - only items from the first planet (StarA) will be used and accounted for.

You can see this if you try to attach Serpulo's TechTree to Planets.sun - no resources from Serpulo will be visible and usable in Serpulo's TechTree in ResearchWindow (or anywhere for that matter).

* Removed spaces

* Added braces, if anyone will need to expand the code later (ig)