try to meet requirements
second draft. based on: 
This doesn't prioritize the sector the player is currently on.
This doesn't pick the closest sector if multiple sectors have the resource.
Erekir specifically has a static loadout, which disregards the universe loadout. You should be using destination.preset.generator.map.rules().loadout as in LaunchLoadoutDialog.